+++ DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/ShisetsuNyutaisho.java
@@ -16,15 +16,13 @@ import jp.co.ndensan.reams.uz.uza.biz.LasdecCode;
 import jp.co.ndensan.reams.uz.uza.biz.ShikibetsuCode;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import jp.co.ndensan.reams.uz.uza.util.db.EntityDataState;
 
 /**
  * 介護保険施設入退所を管理するクラスです。
  */
-public class ShisetsuNyutaisho extends 
-        ModelBase<ShisetsuNyutaishoIdentifier, 
-        DbT1004ShisetsuNyutaishoEntity, ShisetsuNyutaisho> 
+public class ShisetsuNyutaisho extends
+        ModelBase<ShisetsuNyutaishoIdentifier, DbT1004ShisetsuNyutaishoEntity, ShisetsuNyutaisho>
         implements Serializable {
 
     private final DbT1004ShisetsuNyutaishoEntity entity;
@@ -38,7 +36,7 @@ public class ShisetsuNyutaisho extends
      * @param 履歴番号 履歴番号
      */
     public ShisetsuNyutaisho(ShikibetsuCode 識別コード,
-            Decimal 履歴番号) {
+            int 履歴番号) {
         requireNonNull(識別コード, UrSystemErrorMessages.値がnull.getReplacedMessage("識別コード"));
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
         this.entity = new DbT1004ShisetsuNyutaishoEntity();
@@ -92,7 +90,7 @@ public class ShisetsuNyutaisho extends
      *
      * @return 履歴番号
      */
-    public Decimal get履歴番号() {
+    public int get履歴番号() {
         return entity.getRirekiNo();
     }
